requirements.txt
ollama
langfuse
deepeval
pytest
tenacity
pydantic
python-dotenv
